refactor(guide-details): extract my expansion tile
